requirements.txt
streamlit
opencv-python
pytesseract
PyMuPDF
pandas
python-dateutil
requests
